feat: 자동 전송 로직 및 애니메이션 추가 #38
